LayoutTests/editing/pasteboard/drag-files-to-editable-element-as-URLs-expected.txt

If we drag files onto an editable area, then attachments should be inserted into the editable area.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.HTMLAttachmentElement is undefined.
PASS document.createElement("attachment") instanceof HTMLUnknownElement is true
PASS editable.querySelector("attachment") is null
PASS successfullyParsed is true

TEST COMPLETE
